+++ src/typescript/src/domain/errors/DomainErrors.ts
@@ -1,0 +1,14 @@
+export class DomainError extends Error {
+    constructor(message: string) {
+      super(message);
+      this.name = this.constructor.name;
+      Error.captureStackTrace(this, this.constructor);
+    }
+  }
+  
+  export class LampNotFoundError extends DomainError {
+    constructor(id: string) {
+      super(`Lamp with ID ${id} not found`);
+    }
+  }
+
