--- src/typescript/src/domain/errors/lamp-not-found.error.ts
@@ -1,6 +1,0 @@
-export class LampNotFoundError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'LampNotFoundError';
-  }
-}
